Remove highlight untyped from status bar (#7605)
This only changes when the server state changes, which does not always
change when the highlight untyped status is toggled.

Rather than figure out how to get it to not show a stale value, I figure
it's easier to just delete the logic. We have a log line for debugging
if we need to debug it.